backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0

